requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
nostr>=0.9.0
lnbits>=0.3.0
httpx>=0.27.0
